Separate EditProfileE.jsx and EditProfileJS.jsx
